refactor: move rendering to types/xy_chart/rendering
