+++ src/components/react_canvas/bar_geometries.tsx
@@ -3,9 +3,9 @@ import React from 'react';
 import { Group, Rect } from 'react-konva';
 import { animated, Spring } from 'react-spring/renderprops-konva.cjs';
 import { LegendItem } from '../../lib/series/legend';
-import { BarGeometry, getGeometryStyle } from '../../lib/series/rendering';
+import { BarGeometry, getGeometryStyle } from '../../types/xy_chart/rendering/rendering';
 import { SharedGeometryStyle } from '../../utils/themes/theme';
-import { buildBarRenderProps } from '../react_canvas/utils/rendering_props_utils';
+import { buildBarRenderProps } from './utils/rendering_props_utils';
 
 interface BarGeometriesDataProps {
   animated?: boolean;
