+++ src/lib/series/tooltip.ts
@@ -1,5 +1,5 @@
 import { TooltipValue } from '../../types/xy_chart/utils/interactions';
-import { IndexedGeometry } from './rendering';
+import { IndexedGeometry } from '../../types/xy_chart/rendering/rendering';
 import { getColorValuesAsString } from './series';
 import { AxisSpec, BasicSeriesSpec } from './specs';
