+++ src/types/xy_chart/rendering/rendering.areas.test.ts
@@ -1,22 +1,21 @@
 /* eslint @typescript-eslint/no-object-literal-type-assertion: off */
-
-import { computeSeriesDomains } from '../../state/utils';
-import { getGroupId, getSpecId, SpecId } from '../../utils/ids';
-import { ScaleType } from '../../utils/scales/scales';
-import { CurveType } from '../../utils/curves';
-import { IndexedGeometry, LineGeometry, PointGeometry, renderLine } from './rendering';
-import { computeXScale, computeYScales } from './scales';
-import { LineSeriesSpec } from './specs';
-import { LIGHT_THEME } from '../../utils/themes/light_theme';
+import { computeSeriesDomains } from '../../../state/utils';
+import { getGroupId, getSpecId, SpecId } from '../../../utils/ids';
+import { ScaleType } from '../../../utils/scales/scales';
+import { CurveType } from '../../../utils/curves';
+import { AreaGeometry, IndexedGeometry, PointGeometry, renderArea } from './rendering';
+import { computeXScale, computeYScales } from '../../../lib/series/scales';
+import { AreaSeriesSpec } from '../../../lib/series/specs';
+import { LIGHT_THEME } from '../../../utils/themes/light_theme';
 const SPEC_ID = getSpecId('spec_1');
 const GROUP_ID = getGroupId('group_1');
 
-describe('Rendering points - line', () => {
+describe('Rendering points - areas', () => {
   describe('Empty line for missing data', () => {
-    const pointSeriesSpec: LineSeriesSpec = {
+    const pointSeriesSpec: AreaSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -24,18 +23,18 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Ordinal,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 100);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
-    let renderedLine: {
-      lineGeometry: LineGeometry;
+    let renderedArea: {
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedLine = renderLine(
+      renderedArea = renderArea(
         25, // adding a ideal 25px shift, generally applied by renderGeometries
         [],
         xScale,
@@ -46,23 +45,26 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
     });
-    test('Can render the geometry without a line', () => {
-      const { lineGeometry } = renderedLine;
-      expect(lineGeometry.line).toBe('');
-      expect(lineGeometry.color).toBe('red');
-      expect(lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(lineGeometry.geometryId.specId).toEqual(SPEC_ID);
-      expect(lineGeometry.transform).toEqual({ x: 25, y: 0 });
+    test('Render geometry but empty upper and lower lines and area paths', () => {
+      const {
+        areaGeometry: { lines, area, color, geometryId, transform },
+      } = renderedArea;
+      expect(lines.length).toBe(0);
+      expect(area).toBe('');
+      expect(color).toBe('red');
+      expect(geometryId.seriesKey).toEqual([]);
+      expect(geometryId.specId).toEqual(SPEC_ID);
+      expect(transform).toEqual({ x: 25, y: 0 });
     });
   });
-  describe('Single series line chart - ordinal', () => {
-    const pointSeriesSpec: LineSeriesSpec = {
+  describe('Single series area chart - ordinal', () => {
+    const pointSeriesSpec: AreaSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -70,18 +72,18 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Ordinal,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 100);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
-    let renderedLine: {
-      lineGeometry: LineGeometry;
+    let renderedArea: {
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedLine = renderLine(
+      renderedArea = renderArea(
         25, // adding a ideal 25px shift, generally applied by renderGeometries
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -92,22 +94,26 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
     });
-    test('Can render a line', () => {
-      const { lineGeometry } = renderedLine;
-      expect(lineGeometry.line).toBe('M0,0L50,50');
-      expect(lineGeometry.color).toBe('red');
-      expect(lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(lineGeometry.geometryId.specId).toEqual(SPEC_ID);
-      expect(lineGeometry.transform).toEqual({ x: 25, y: 0 });
+    test('Can render an line and area paths', () => {
+      const {
+        areaGeometry: { lines, area, color, geometryId, transform },
+      } = renderedArea;
+      expect(lines[0]).toBe('M0,0L50,50');
+      expect(area).toBe('M0,0L50,50L50,100L0,100Z');
+      expect(color).toBe('red');
+      expect(geometryId.seriesKey).toEqual([]);
+      expect(geometryId.specId).toEqual(SPEC_ID);
+      expect(transform).toEqual({ x: 25, y: 0 });
     });
+
     test('Can render two points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
-      } = renderedLine;
+      } = renderedArea;
 
       expect(points[0]).toEqual({
         x: 0,
@@ -123,6 +129,7 @@ describe('Rendering points - line', () => {
           x: 0,
           y: 10,
         },
+
         transform: {
           x: 25,
           y: 0,
@@ -150,13 +157,13 @@ describe('Rendering points - line', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Multi series line chart - ordinal', () => {
+  describe('Multi series area chart - ordinal', () => {
     const spec1Id = getSpecId('point1');
     const spec2Id = getSpecId('point2');
-    const pointSeriesSpec1: LineSeriesSpec = {
+    const pointSeriesSpec1: AreaSeriesSpec = {
       id: spec1Id,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -164,10 +171,10 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Ordinal,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesSpec2: LineSeriesSpec = {
+    const pointSeriesSpec2: AreaSeriesSpec = {
       id: spec2Id,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[0, 20], [1, 10]],
       xAccessor: 0,
@@ -175,7 +182,7 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Ordinal,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
     pointSeriesMap.set(spec1Id, pointSeriesSpec1);
     pointSeriesMap.set(spec2Id, pointSeriesSpec2);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
@@ -183,16 +190,16 @@ describe('Rendering points - line', () => {
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
     let firstLine: {
-      lineGeometry: LineGeometry;
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
     let secondLine: {
-      lineGeometry: LineGeometry;
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      firstLine = renderLine(
+      firstLine = renderArea(
         25, // adding a ideal 25px shift, generally applied by renderGeometries
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -203,9 +210,9 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
-      secondLine = renderLine(
+      secondLine = renderArea(
         25, // adding a ideal 25px shift, generally applied by renderGeometries
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[1].data,
         xScale,
@@ -216,26 +223,28 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
     });
 
-    test('Can render two ordinal lines', () => {
-      expect(firstLine.lineGeometry.line).toBe('M0,50L50,75');
-      expect(firstLine.lineGeometry.color).toBe('red');
-      expect(firstLine.lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(firstLine.lineGeometry.geometryId.specId).toEqual(spec1Id);
-      expect(firstLine.lineGeometry.transform).toEqual({ x: 25, y: 0 });
+    test('Can render two ordinal areas', () => {
+      expect(firstLine.areaGeometry.lines[0]).toBe('M0,50L50,75');
+      expect(firstLine.areaGeometry.area).toBe('M0,50L50,75L50,100L0,100Z');
+      expect(firstLine.areaGeometry.color).toBe('red');
+      expect(firstLine.areaGeometry.geometryId.seriesKey).toEqual([]);
+      expect(firstLine.areaGeometry.geometryId.specId).toEqual(spec1Id);
+      expect(firstLine.areaGeometry.transform).toEqual({ x: 25, y: 0 });
 
-      expect(secondLine.lineGeometry.line).toBe('M0,0L50,50');
-      expect(secondLine.lineGeometry.color).toBe('blue');
-      expect(secondLine.lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(secondLine.lineGeometry.geometryId.specId).toEqual(spec2Id);
-      expect(secondLine.lineGeometry.transform).toEqual({ x: 25, y: 0 });
+      expect(secondLine.areaGeometry.lines[0]).toBe('M0,0L50,50');
+      expect(secondLine.areaGeometry.area).toBe('M0,0L50,50L50,100L0,100Z');
+      expect(secondLine.areaGeometry.color).toBe('blue');
+      expect(secondLine.areaGeometry.geometryId.seriesKey).toEqual([]);
+      expect(secondLine.areaGeometry.geometryId.specId).toEqual(spec2Id);
+      expect(secondLine.areaGeometry.transform).toEqual({ x: 25, y: 0 });
     });
     test('can render first spec points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
       } = firstLine;
       expect(points.length).toEqual(2);
@@ -261,8 +270,8 @@ describe('Rendering points - line', () => {
       expect(points[1]).toEqual({
         x: 50,
         y: 75,
-        color: 'red',
         radius: 10,
+        color: 'red',
         geometryId: {
           specId: spec1Id,
           seriesKey: [],
@@ -281,15 +290,15 @@ describe('Rendering points - line', () => {
     });
     test('can render second spec points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
       } = secondLine;
       expect(points.length).toEqual(2);
       expect(points[0]).toEqual({
         x: 0,
         y: 0,
-        color: 'blue',
         radius: 10,
+        color: 'blue',
         geometryId: {
           specId: spec2Id,
           seriesKey: [],
@@ -307,8 +316,8 @@ describe('Rendering points - line', () => {
       expect(points[1]).toEqual({
         x: 50,
         y: 50,
-        color: 'blue',
         radius: 10,
+        color: 'blue',
         geometryId: {
           specId: spec2Id,
           seriesKey: [],
@@ -326,11 +335,11 @@ describe('Rendering points - line', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Single series line chart - linear', () => {
-    const pointSeriesSpec: LineSeriesSpec = {
+  describe('Single series area chart - linear', () => {
+    const pointSeriesSpec: AreaSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -338,19 +347,19 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Linear,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 100);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
-    let renderedLine: {
-      lineGeometry: LineGeometry;
+    let renderedArea: {
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedLine = renderLine(
+      renderedArea = renderArea(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -361,26 +370,27 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
     });
-    test('Can render a linear line', () => {
-      expect(renderedLine.lineGeometry.line).toBe('M0,0L100,50');
-      expect(renderedLine.lineGeometry.color).toBe('red');
-      expect(renderedLine.lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(renderedLine.lineGeometry.geometryId.specId).toEqual(SPEC_ID);
-      expect(renderedLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
+    test('Can render a linear area', () => {
+      expect(renderedArea.areaGeometry.lines[0]).toBe('M0,0L100,50');
+      expect(renderedArea.areaGeometry.area).toBe('M0,0L100,50L100,100L0,100Z');
+      expect(renderedArea.areaGeometry.color).toBe('red');
+      expect(renderedArea.areaGeometry.geometryId.seriesKey).toEqual([]);
+      expect(renderedArea.areaGeometry.geometryId.specId).toEqual(SPEC_ID);
+      expect(renderedArea.areaGeometry.transform).toEqual({ x: 0, y: 0 });
     });
     test('Can render two points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
-      } = renderedLine;
+      } = renderedArea;
       expect(points[0]).toEqual({
         x: 0,
         y: 0,
-        color: 'red',
         radius: 10,
+        color: 'red',
         geometryId: {
           specId: SPEC_ID,
           seriesKey: [],
@@ -398,8 +408,8 @@ describe('Rendering points - line', () => {
       expect(points[1]).toEqual({
         x: 100,
         y: 50,
-        color: 'red',
         radius: 10,
+        color: 'red',
         geometryId: {
           specId: SPEC_ID,
           seriesKey: [],
@@ -417,13 +427,13 @@ describe('Rendering points - line', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Multi series line chart - linear', () => {
+  describe('Multi series area chart - linear', () => {
     const spec1Id = getSpecId('point1');
     const spec2Id = getSpecId('point2');
-    const pointSeriesSpec1: LineSeriesSpec = {
+    const pointSeriesSpec1: AreaSeriesSpec = {
       id: spec1Id,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -431,10 +441,10 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Linear,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesSpec2: LineSeriesSpec = {
+    const pointSeriesSpec2: AreaSeriesSpec = {
       id: spec2Id,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[0, 20], [1, 10]],
       xAccessor: 0,
@@ -442,7 +452,7 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Linear,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
     pointSeriesMap.set(spec1Id, pointSeriesSpec1);
     pointSeriesMap.set(spec2Id, pointSeriesSpec2);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
@@ -450,16 +460,16 @@ describe('Rendering points - line', () => {
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
     let firstLine: {
-      lineGeometry: LineGeometry;
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
     let secondLine: {
-      lineGeometry: LineGeometry;
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      firstLine = renderLine(
+      firstLine = renderArea(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -470,9 +480,9 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
-      secondLine = renderLine(
+      secondLine = renderArea(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[1].data,
         xScale,
@@ -483,25 +493,27 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
     });
-    test('can render two linear lines', () => {
-      expect(firstLine.lineGeometry.line).toBe('M0,50L100,75');
-      expect(firstLine.lineGeometry.color).toBe('red');
-      expect(firstLine.lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(firstLine.lineGeometry.geometryId.specId).toEqual(spec1Id);
-      expect(firstLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
+    test('can render two linear areas', () => {
+      expect(firstLine.areaGeometry.lines[0]).toBe('M0,50L100,75');
+      expect(firstLine.areaGeometry.area).toBe('M0,50L100,75L100,100L0,100Z');
+      expect(firstLine.areaGeometry.color).toBe('red');
+      expect(firstLine.areaGeometry.geometryId.seriesKey).toEqual([]);
+      expect(firstLine.areaGeometry.geometryId.specId).toEqual(spec1Id);
+      expect(firstLine.areaGeometry.transform).toEqual({ x: 0, y: 0 });
 
-      expect(secondLine.lineGeometry.line).toBe('M0,0L100,50');
-      expect(secondLine.lineGeometry.color).toBe('blue');
-      expect(secondLine.lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(secondLine.lineGeometry.geometryId.specId).toEqual(spec2Id);
-      expect(secondLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
+      expect(secondLine.areaGeometry.lines[0]).toBe('M0,0L100,50');
+      expect(secondLine.areaGeometry.area).toBe('M0,0L100,50L100,100L0,100Z');
+      expect(secondLine.areaGeometry.color).toBe('blue');
+      expect(secondLine.areaGeometry.geometryId.seriesKey).toEqual([]);
+      expect(secondLine.areaGeometry.geometryId.specId).toEqual(spec2Id);
+      expect(secondLine.areaGeometry.transform).toEqual({ x: 0, y: 0 });
     });
     test('can render first spec points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
       } = firstLine;
       expect(points.length).toEqual(2);
@@ -547,15 +559,15 @@ describe('Rendering points - line', () => {
     });
     test('can render second spec points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
       } = secondLine;
       expect(points.length).toEqual(2);
       expect(points[0]).toEqual({
         x: 0,
         y: 0,
-        color: 'blue',
         radius: 10,
+        color: 'blue',
         geometryId: {
           specId: spec2Id,
           seriesKey: [],
@@ -573,8 +585,8 @@ describe('Rendering points - line', () => {
       expect(points[1]).toEqual({
         x: 100,
         y: 50,
-        color: 'blue',
         radius: 10,
+        color: 'blue',
         geometryId: {
           specId: spec2Id,
           seriesKey: [],
@@ -592,11 +604,11 @@ describe('Rendering points - line', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Single series line chart - time', () => {
-    const pointSeriesSpec: LineSeriesSpec = {
+  describe('Single series area chart - time', () => {
+    const pointSeriesSpec: AreaSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[1546300800000, 10], [1546387200000, 5]],
       xAccessor: 0,
@@ -604,19 +616,19 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Time,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 100);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
-    let renderedLine: {
-      lineGeometry: LineGeometry;
+    let renderedArea: {
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedLine = renderLine(
+      renderedArea = renderArea(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -627,21 +639,22 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
     });
-    test('Can render a time line', () => {
-      expect(renderedLine.lineGeometry.line).toBe('M0,0L100,50');
-      expect(renderedLine.lineGeometry.color).toBe('red');
-      expect(renderedLine.lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(renderedLine.lineGeometry.geometryId.specId).toEqual(SPEC_ID);
-      expect(renderedLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
+    test('Can render a time area', () => {
+      expect(renderedArea.areaGeometry.lines[0]).toBe('M0,0L100,50');
+      expect(renderedArea.areaGeometry.area).toBe('M0,0L100,50L100,100L0,100Z');
+      expect(renderedArea.areaGeometry.color).toBe('red');
+      expect(renderedArea.areaGeometry.geometryId.seriesKey).toEqual([]);
+      expect(renderedArea.areaGeometry.geometryId.specId).toEqual(SPEC_ID);
+      expect(renderedArea.areaGeometry.transform).toEqual({ x: 0, y: 0 });
     });
     test('Can render two points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
-      } = renderedLine;
+      } = renderedArea;
       expect(points[0]).toEqual({
         x: 0,
         y: 0,
@@ -683,13 +696,13 @@ describe('Rendering points - line', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Multi series line chart - time', () => {
+  describe('Multi series area chart - time', () => {
     const spec1Id = getSpecId('point1');
     const spec2Id = getSpecId('point2');
-    const pointSeriesSpec1: LineSeriesSpec = {
+    const pointSeriesSpec1: AreaSeriesSpec = {
       id: spec1Id,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[1546300800000, 10], [1546387200000, 5]],
       xAccessor: 0,
@@ -697,10 +710,10 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Time,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesSpec2: LineSeriesSpec = {
+    const pointSeriesSpec2: AreaSeriesSpec = {
       id: spec2Id,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[1546300800000, 20], [1546387200000, 10]],
       xAccessor: 0,
@@ -708,7 +721,7 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Time,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
     pointSeriesMap.set(spec1Id, pointSeriesSpec1);
     pointSeriesMap.set(spec2Id, pointSeriesSpec2);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
@@ -716,16 +729,16 @@ describe('Rendering points - line', () => {
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
     let firstLine: {
-      lineGeometry: LineGeometry;
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
     let secondLine: {
-      lineGeometry: LineGeometry;
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      firstLine = renderLine(
+      firstLine = renderArea(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -736,9 +749,9 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
-      secondLine = renderLine(
+      secondLine = renderArea(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[1].data,
         xScale,
@@ -749,12 +762,12 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
     });
     test('can render first spec points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
       } = firstLine;
       expect(points.length).toEqual(2);
@@ -800,7 +813,7 @@ describe('Rendering points - line', () => {
     });
     test('can render second spec points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
       } = secondLine;
       expect(points.length).toEqual(2);
@@ -845,11 +858,11 @@ describe('Rendering points - line', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Single series line chart - y log', () => {
-    const pointSeriesSpec: LineSeriesSpec = {
+  describe('Single series area chart - y log', () => {
+    const pointSeriesSpec: AreaSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'line',
+      seriesType: 'area',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5], [2, null], [3, 5], [4, 5], [5, 0], [6, 10], [7, 10], [8, 10]],
       xAccessor: 0,
@@ -857,19 +870,19 @@ describe('Rendering points - line', () => {
       xScaleType: ScaleType.Linear,
       yScaleType: ScaleType.Log,
     };
-    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 90);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
-    let renderedLine: {
-      lineGeometry: LineGeometry;
+    let renderedArea: {
+      areaGeometry: AreaGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedLine = renderLine(
+      renderedArea = renderArea(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -880,22 +893,23 @@ describe('Rendering points - line', () => {
         false,
         [],
         0,
-        LIGHT_THEME.lineSeriesStyle,
+        LIGHT_THEME.areaSeriesStyle,
       );
     });
-    test('Can render a splitted line', () => {
-      // expect(renderedLine.lineGeometry.line).toBe('ss');
-      expect(renderedLine.lineGeometry.line.split('M').length - 1).toBe(3);
-      expect(renderedLine.lineGeometry.color).toBe('red');
-      expect(renderedLine.lineGeometry.geometryId.seriesKey).toEqual([]);
-      expect(renderedLine.lineGeometry.geometryId.specId).toEqual(SPEC_ID);
-      expect(renderedLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
+    test('Can render a splitted area and line', () => {
+      // expect(renderedArea.lineGeometry.line).toBe('ss');
+      expect(renderedArea.areaGeometry.lines[0].split('M').length - 1).toBe(3);
+      expect(renderedArea.areaGeometry.area.split('M').length - 1).toBe(3);
+      expect(renderedArea.areaGeometry.color).toBe('red');
+      expect(renderedArea.areaGeometry.geometryId.seriesKey).toEqual([]);
+      expect(renderedArea.areaGeometry.geometryId.specId).toEqual(SPEC_ID);
+      expect(renderedArea.areaGeometry.transform).toEqual({ x: 0, y: 0 });
     });
-    test('Can render points', () => {
+    test('Can render points points', () => {
       const {
-        lineGeometry: { points },
+        areaGeometry: { points },
         indexedGeometries,
-      } = renderedLine;
+      } = renderedArea;
       // all the points minus the undefined ones on a log scale
       expect(points.length).toBe(7);
       // all the points expect null geometries
@@ -907,7 +921,7 @@ describe('Rendering points - line', () => {
       expect(zeroValueIndexdGeometry).toBeDefined();
       expect(zeroValueIndexdGeometry.length).toBe(1);
       // moved to the bottom of the chart
-      expect((zeroValueIndexdGeometry[0] as PointGeometry).y).toBe(100);
+      expect(zeroValueIndexdGeometry[0].y).toBe(100);
       // 0 radius point
       expect((zeroValueIndexdGeometry[0] as PointGeometry).radius).toBe(0);
     });
