+++ src/types/xy_chart/rendering/rendering.lines.test.ts
@@ -1,21 +1,23 @@
 /* eslint @typescript-eslint/no-object-literal-type-assertion: off */
-import { computeSeriesDomains } from '../../state/utils';
-import { getGroupId, getSpecId, SpecId } from '../../utils/ids';
-import { ScaleType } from '../../utils/scales/scales';
-import { CurveType } from '../../utils/curves';
-import { AreaGeometry, IndexedGeometry, PointGeometry, renderArea } from './rendering';
-import { computeXScale, computeYScales } from './scales';
-import { AreaSeriesSpec } from './specs';
-import { LIGHT_THEME } from '../../utils/themes/light_theme';
+
+import { computeSeriesDomains } from '../../../state/utils';
+import { getGroupId, getSpecId, SpecId } from '../../../utils/ids';
+import { ScaleType } from '../../../utils/scales/scales';
+import { CurveType } from '../../../utils/curves';
+import { IndexedGeometry, LineGeometry, PointGeometry, renderLine } from './rendering';
+import { computeXScale, computeYScales } from '../../../lib/series/scales';
+import { LineSeriesSpec } from '../../../lib/series/specs';
+import { LIGHT_THEME } from '../../../utils/themes/light_theme';
+
 const SPEC_ID = getSpecId('spec_1');
 const GROUP_ID = getGroupId('group_1');
 
-describe('Rendering points - areas', () => {
+describe('Rendering points - line', () => {
   describe('Empty line for missing data', () => {
-    const pointSeriesSpec: AreaSeriesSpec = {
+    const pointSeriesSpec: LineSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -23,18 +25,18 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Ordinal,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 100);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
-    let renderedArea: {
-      areaGeometry: AreaGeometry;
+    let renderedLine: {
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedArea = renderArea(
+      renderedLine = renderLine(
         25, // adding a ideal 25px shift, generally applied by renderGeometries
         [],
         xScale,
@@ -45,26 +47,23 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
     });
-    test('Render geometry but empty upper and lower lines and area paths', () => {
-      const {
-        areaGeometry: { lines, area, color, geometryId, transform },
-      } = renderedArea;
-      expect(lines.length).toBe(0);
-      expect(area).toBe('');
-      expect(color).toBe('red');
-      expect(geometryId.seriesKey).toEqual([]);
-      expect(geometryId.specId).toEqual(SPEC_ID);
-      expect(transform).toEqual({ x: 25, y: 0 });
+    test('Can render the geometry without a line', () => {
+      const { lineGeometry } = renderedLine;
+      expect(lineGeometry.line).toBe('');
+      expect(lineGeometry.color).toBe('red');
+      expect(lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(lineGeometry.geometryId.specId).toEqual(SPEC_ID);
+      expect(lineGeometry.transform).toEqual({ x: 25, y: 0 });
     });
   });
-  describe('Single series area chart - ordinal', () => {
-    const pointSeriesSpec: AreaSeriesSpec = {
+  describe('Single series line chart - ordinal', () => {
+    const pointSeriesSpec: LineSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -72,18 +71,18 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Ordinal,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 100);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
-    let renderedArea: {
-      areaGeometry: AreaGeometry;
+    let renderedLine: {
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedArea = renderArea(
+      renderedLine = renderLine(
         25, // adding a ideal 25px shift, generally applied by renderGeometries
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -94,26 +93,22 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
     });
-    test('Can render an line and area paths', () => {
-      const {
-        areaGeometry: { lines, area, color, geometryId, transform },
-      } = renderedArea;
-      expect(lines[0]).toBe('M0,0L50,50');
-      expect(area).toBe('M0,0L50,50L50,100L0,100Z');
-      expect(color).toBe('red');
-      expect(geometryId.seriesKey).toEqual([]);
-      expect(geometryId.specId).toEqual(SPEC_ID);
-      expect(transform).toEqual({ x: 25, y: 0 });
+    test('Can render a line', () => {
+      const { lineGeometry } = renderedLine;
+      expect(lineGeometry.line).toBe('M0,0L50,50');
+      expect(lineGeometry.color).toBe('red');
+      expect(lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(lineGeometry.geometryId.specId).toEqual(SPEC_ID);
+      expect(lineGeometry.transform).toEqual({ x: 25, y: 0 });
     });
-
     test('Can render two points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
-      } = renderedArea;
+      } = renderedLine;
 
       expect(points[0]).toEqual({
         x: 0,
@@ -129,7 +124,6 @@ describe('Rendering points - areas', () => {
           x: 0,
           y: 10,
         },
-
         transform: {
           x: 25,
           y: 0,
@@ -157,13 +151,13 @@ describe('Rendering points - areas', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Multi series area chart - ordinal', () => {
+  describe('Multi series line chart - ordinal', () => {
     const spec1Id = getSpecId('point1');
     const spec2Id = getSpecId('point2');
-    const pointSeriesSpec1: AreaSeriesSpec = {
+    const pointSeriesSpec1: LineSeriesSpec = {
       id: spec1Id,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -171,10 +165,10 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Ordinal,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesSpec2: AreaSeriesSpec = {
+    const pointSeriesSpec2: LineSeriesSpec = {
       id: spec2Id,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[0, 20], [1, 10]],
       xAccessor: 0,
@@ -182,7 +176,7 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Ordinal,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
     pointSeriesMap.set(spec1Id, pointSeriesSpec1);
     pointSeriesMap.set(spec2Id, pointSeriesSpec2);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
@@ -190,16 +184,16 @@ describe('Rendering points - areas', () => {
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
     let firstLine: {
-      areaGeometry: AreaGeometry;
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
     let secondLine: {
-      areaGeometry: AreaGeometry;
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      firstLine = renderArea(
+      firstLine = renderLine(
         25, // adding a ideal 25px shift, generally applied by renderGeometries
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -210,9 +204,9 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
-      secondLine = renderArea(
+      secondLine = renderLine(
         25, // adding a ideal 25px shift, generally applied by renderGeometries
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[1].data,
         xScale,
@@ -223,28 +217,26 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
     });
 
-    test('Can render two ordinal areas', () => {
-      expect(firstLine.areaGeometry.lines[0]).toBe('M0,50L50,75');
-      expect(firstLine.areaGeometry.area).toBe('M0,50L50,75L50,100L0,100Z');
-      expect(firstLine.areaGeometry.color).toBe('red');
-      expect(firstLine.areaGeometry.geometryId.seriesKey).toEqual([]);
-      expect(firstLine.areaGeometry.geometryId.specId).toEqual(spec1Id);
-      expect(firstLine.areaGeometry.transform).toEqual({ x: 25, y: 0 });
+    test('Can render two ordinal lines', () => {
+      expect(firstLine.lineGeometry.line).toBe('M0,50L50,75');
+      expect(firstLine.lineGeometry.color).toBe('red');
+      expect(firstLine.lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(firstLine.lineGeometry.geometryId.specId).toEqual(spec1Id);
+      expect(firstLine.lineGeometry.transform).toEqual({ x: 25, y: 0 });
 
-      expect(secondLine.areaGeometry.lines[0]).toBe('M0,0L50,50');
-      expect(secondLine.areaGeometry.area).toBe('M0,0L50,50L50,100L0,100Z');
-      expect(secondLine.areaGeometry.color).toBe('blue');
-      expect(secondLine.areaGeometry.geometryId.seriesKey).toEqual([]);
-      expect(secondLine.areaGeometry.geometryId.specId).toEqual(spec2Id);
-      expect(secondLine.areaGeometry.transform).toEqual({ x: 25, y: 0 });
+      expect(secondLine.lineGeometry.line).toBe('M0,0L50,50');
+      expect(secondLine.lineGeometry.color).toBe('blue');
+      expect(secondLine.lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(secondLine.lineGeometry.geometryId.specId).toEqual(spec2Id);
+      expect(secondLine.lineGeometry.transform).toEqual({ x: 25, y: 0 });
     });
     test('can render first spec points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
       } = firstLine;
       expect(points.length).toEqual(2);
@@ -270,8 +262,8 @@ describe('Rendering points - areas', () => {
       expect(points[1]).toEqual({
         x: 50,
         y: 75,
-        radius: 10,
         color: 'red',
+        radius: 10,
         geometryId: {
           specId: spec1Id,
           seriesKey: [],
@@ -290,15 +282,15 @@ describe('Rendering points - areas', () => {
     });
     test('can render second spec points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
       } = secondLine;
       expect(points.length).toEqual(2);
       expect(points[0]).toEqual({
         x: 0,
         y: 0,
-        radius: 10,
         color: 'blue',
+        radius: 10,
         geometryId: {
           specId: spec2Id,
           seriesKey: [],
@@ -316,8 +308,8 @@ describe('Rendering points - areas', () => {
       expect(points[1]).toEqual({
         x: 50,
         y: 50,
-        radius: 10,
         color: 'blue',
+        radius: 10,
         geometryId: {
           specId: spec2Id,
           seriesKey: [],
@@ -335,11 +327,11 @@ describe('Rendering points - areas', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Single series area chart - linear', () => {
-    const pointSeriesSpec: AreaSeriesSpec = {
+  describe('Single series line chart - linear', () => {
+    const pointSeriesSpec: LineSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -347,19 +339,19 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Linear,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 100);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
-    let renderedArea: {
-      areaGeometry: AreaGeometry;
+    let renderedLine: {
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedArea = renderArea(
+      renderedLine = renderLine(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -370,27 +362,26 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
     });
-    test('Can render a linear area', () => {
-      expect(renderedArea.areaGeometry.lines[0]).toBe('M0,0L100,50');
-      expect(renderedArea.areaGeometry.area).toBe('M0,0L100,50L100,100L0,100Z');
-      expect(renderedArea.areaGeometry.color).toBe('red');
-      expect(renderedArea.areaGeometry.geometryId.seriesKey).toEqual([]);
-      expect(renderedArea.areaGeometry.geometryId.specId).toEqual(SPEC_ID);
-      expect(renderedArea.areaGeometry.transform).toEqual({ x: 0, y: 0 });
+    test('Can render a linear line', () => {
+      expect(renderedLine.lineGeometry.line).toBe('M0,0L100,50');
+      expect(renderedLine.lineGeometry.color).toBe('red');
+      expect(renderedLine.lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(renderedLine.lineGeometry.geometryId.specId).toEqual(SPEC_ID);
+      expect(renderedLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
     });
     test('Can render two points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
-      } = renderedArea;
+      } = renderedLine;
       expect(points[0]).toEqual({
         x: 0,
         y: 0,
-        radius: 10,
         color: 'red',
+        radius: 10,
         geometryId: {
           specId: SPEC_ID,
           seriesKey: [],
@@ -408,8 +399,8 @@ describe('Rendering points - areas', () => {
       expect(points[1]).toEqual({
         x: 100,
         y: 50,
-        radius: 10,
         color: 'red',
+        radius: 10,
         geometryId: {
           specId: SPEC_ID,
           seriesKey: [],
@@ -427,13 +418,13 @@ describe('Rendering points - areas', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Multi series area chart - linear', () => {
+  describe('Multi series line chart - linear', () => {
     const spec1Id = getSpecId('point1');
     const spec2Id = getSpecId('point2');
-    const pointSeriesSpec1: AreaSeriesSpec = {
+    const pointSeriesSpec1: LineSeriesSpec = {
       id: spec1Id,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5]],
       xAccessor: 0,
@@ -441,10 +432,10 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Linear,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesSpec2: AreaSeriesSpec = {
+    const pointSeriesSpec2: LineSeriesSpec = {
       id: spec2Id,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[0, 20], [1, 10]],
       xAccessor: 0,
@@ -452,7 +443,7 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Linear,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
     pointSeriesMap.set(spec1Id, pointSeriesSpec1);
     pointSeriesMap.set(spec2Id, pointSeriesSpec2);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
@@ -460,16 +451,16 @@ describe('Rendering points - areas', () => {
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
     let firstLine: {
-      areaGeometry: AreaGeometry;
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
     let secondLine: {
-      areaGeometry: AreaGeometry;
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      firstLine = renderArea(
+      firstLine = renderLine(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -480,9 +471,9 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
-      secondLine = renderArea(
+      secondLine = renderLine(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[1].data,
         xScale,
@@ -493,27 +484,25 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
     });
-    test('can render two linear areas', () => {
-      expect(firstLine.areaGeometry.lines[0]).toBe('M0,50L100,75');
-      expect(firstLine.areaGeometry.area).toBe('M0,50L100,75L100,100L0,100Z');
-      expect(firstLine.areaGeometry.color).toBe('red');
-      expect(firstLine.areaGeometry.geometryId.seriesKey).toEqual([]);
-      expect(firstLine.areaGeometry.geometryId.specId).toEqual(spec1Id);
-      expect(firstLine.areaGeometry.transform).toEqual({ x: 0, y: 0 });
+    test('can render two linear lines', () => {
+      expect(firstLine.lineGeometry.line).toBe('M0,50L100,75');
+      expect(firstLine.lineGeometry.color).toBe('red');
+      expect(firstLine.lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(firstLine.lineGeometry.geometryId.specId).toEqual(spec1Id);
+      expect(firstLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
 
-      expect(secondLine.areaGeometry.lines[0]).toBe('M0,0L100,50');
-      expect(secondLine.areaGeometry.area).toBe('M0,0L100,50L100,100L0,100Z');
-      expect(secondLine.areaGeometry.color).toBe('blue');
-      expect(secondLine.areaGeometry.geometryId.seriesKey).toEqual([]);
-      expect(secondLine.areaGeometry.geometryId.specId).toEqual(spec2Id);
-      expect(secondLine.areaGeometry.transform).toEqual({ x: 0, y: 0 });
+      expect(secondLine.lineGeometry.line).toBe('M0,0L100,50');
+      expect(secondLine.lineGeometry.color).toBe('blue');
+      expect(secondLine.lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(secondLine.lineGeometry.geometryId.specId).toEqual(spec2Id);
+      expect(secondLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
     });
     test('can render first spec points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
       } = firstLine;
       expect(points.length).toEqual(2);
@@ -559,15 +548,15 @@ describe('Rendering points - areas', () => {
     });
     test('can render second spec points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
       } = secondLine;
       expect(points.length).toEqual(2);
       expect(points[0]).toEqual({
         x: 0,
         y: 0,
-        radius: 10,
         color: 'blue',
+        radius: 10,
         geometryId: {
           specId: spec2Id,
           seriesKey: [],
@@ -585,8 +574,8 @@ describe('Rendering points - areas', () => {
       expect(points[1]).toEqual({
         x: 100,
         y: 50,
-        radius: 10,
         color: 'blue',
+        radius: 10,
         geometryId: {
           specId: spec2Id,
           seriesKey: [],
@@ -604,11 +593,11 @@ describe('Rendering points - areas', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Single series area chart - time', () => {
-    const pointSeriesSpec: AreaSeriesSpec = {
+  describe('Single series line chart - time', () => {
+    const pointSeriesSpec: LineSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[1546300800000, 10], [1546387200000, 5]],
       xAccessor: 0,
@@ -616,19 +605,19 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Time,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 100);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
-    let renderedArea: {
-      areaGeometry: AreaGeometry;
+    let renderedLine: {
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedArea = renderArea(
+      renderedLine = renderLine(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -639,22 +628,21 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
     });
-    test('Can render a time area', () => {
-      expect(renderedArea.areaGeometry.lines[0]).toBe('M0,0L100,50');
-      expect(renderedArea.areaGeometry.area).toBe('M0,0L100,50L100,100L0,100Z');
-      expect(renderedArea.areaGeometry.color).toBe('red');
-      expect(renderedArea.areaGeometry.geometryId.seriesKey).toEqual([]);
-      expect(renderedArea.areaGeometry.geometryId.specId).toEqual(SPEC_ID);
-      expect(renderedArea.areaGeometry.transform).toEqual({ x: 0, y: 0 });
+    test('Can render a time line', () => {
+      expect(renderedLine.lineGeometry.line).toBe('M0,0L100,50');
+      expect(renderedLine.lineGeometry.color).toBe('red');
+      expect(renderedLine.lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(renderedLine.lineGeometry.geometryId.specId).toEqual(SPEC_ID);
+      expect(renderedLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
     });
     test('Can render two points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
-      } = renderedArea;
+      } = renderedLine;
       expect(points[0]).toEqual({
         x: 0,
         y: 0,
@@ -696,13 +684,13 @@ describe('Rendering points - areas', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Multi series area chart - time', () => {
+  describe('Multi series line chart - time', () => {
     const spec1Id = getSpecId('point1');
     const spec2Id = getSpecId('point2');
-    const pointSeriesSpec1: AreaSeriesSpec = {
+    const pointSeriesSpec1: LineSeriesSpec = {
       id: spec1Id,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[1546300800000, 10], [1546387200000, 5]],
       xAccessor: 0,
@@ -710,10 +698,10 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Time,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesSpec2: AreaSeriesSpec = {
+    const pointSeriesSpec2: LineSeriesSpec = {
       id: spec2Id,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[1546300800000, 20], [1546387200000, 10]],
       xAccessor: 0,
@@ -721,7 +709,7 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Time,
       yScaleType: ScaleType.Linear,
     };
-    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
     pointSeriesMap.set(spec1Id, pointSeriesSpec1);
     pointSeriesMap.set(spec2Id, pointSeriesSpec2);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
@@ -729,16 +717,16 @@ describe('Rendering points - areas', () => {
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
     let firstLine: {
-      areaGeometry: AreaGeometry;
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
     let secondLine: {
-      areaGeometry: AreaGeometry;
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      firstLine = renderArea(
+      firstLine = renderLine(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -749,9 +737,9 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
-      secondLine = renderArea(
+      secondLine = renderLine(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[1].data,
         xScale,
@@ -762,12 +750,12 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
     });
     test('can render first spec points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
       } = firstLine;
       expect(points.length).toEqual(2);
@@ -813,7 +801,7 @@ describe('Rendering points - areas', () => {
     });
     test('can render second spec points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
       } = secondLine;
       expect(points.length).toEqual(2);
@@ -858,11 +846,11 @@ describe('Rendering points - areas', () => {
       expect(indexedGeometries.size).toEqual(points.length);
     });
   });
-  describe('Single series area chart - y log', () => {
-    const pointSeriesSpec: AreaSeriesSpec = {
+  describe('Single series line chart - y log', () => {
+    const pointSeriesSpec: LineSeriesSpec = {
       id: SPEC_ID,
       groupId: GROUP_ID,
-      seriesType: 'area',
+      seriesType: 'line',
       yScaleToDataExtent: false,
       data: [[0, 10], [1, 5], [2, null], [3, 5], [4, 5], [5, 0], [6, 10], [7, 10], [8, 10]],
       xAccessor: 0,
@@ -870,19 +858,19 @@ describe('Rendering points - areas', () => {
       xScaleType: ScaleType.Linear,
       yScaleType: ScaleType.Log,
     };
-    const pointSeriesMap = new Map<SpecId, AreaSeriesSpec>();
+    const pointSeriesMap = new Map<SpecId, LineSeriesSpec>();
     pointSeriesMap.set(SPEC_ID, pointSeriesSpec);
     const pointSeriesDomains = computeSeriesDomains(pointSeriesMap, new Map());
     const xScale = computeXScale(pointSeriesDomains.xDomain, pointSeriesMap.size, 0, 90);
     const yScales = computeYScales(pointSeriesDomains.yDomain, 100, 0);
 
-    let renderedArea: {
-      areaGeometry: AreaGeometry;
+    let renderedLine: {
+      lineGeometry: LineGeometry;
       indexedGeometries: Map<any, IndexedGeometry[]>;
     };
 
     beforeEach(() => {
-      renderedArea = renderArea(
+      renderedLine = renderLine(
         0, // not applied any shift, renderGeometries applies it only with mixed charts
         pointSeriesDomains.formattedDataSeries.nonStacked[0].dataSeries[0].data,
         xScale,
@@ -893,23 +881,22 @@ describe('Rendering points - areas', () => {
         false,
         [],
         0,
-        LIGHT_THEME.areaSeriesStyle,
+        LIGHT_THEME.lineSeriesStyle,
       );
     });
-    test('Can render a splitted area and line', () => {
-      // expect(renderedArea.lineGeometry.line).toBe('ss');
-      expect(renderedArea.areaGeometry.lines[0].split('M').length - 1).toBe(3);
-      expect(renderedArea.areaGeometry.area.split('M').length - 1).toBe(3);
-      expect(renderedArea.areaGeometry.color).toBe('red');
-      expect(renderedArea.areaGeometry.geometryId.seriesKey).toEqual([]);
-      expect(renderedArea.areaGeometry.geometryId.specId).toEqual(SPEC_ID);
-      expect(renderedArea.areaGeometry.transform).toEqual({ x: 0, y: 0 });
+    test('Can render a splitted line', () => {
+      // expect(renderedLine.lineGeometry.line).toBe('ss');
+      expect(renderedLine.lineGeometry.line.split('M').length - 1).toBe(3);
+      expect(renderedLine.lineGeometry.color).toBe('red');
+      expect(renderedLine.lineGeometry.geometryId.seriesKey).toEqual([]);
+      expect(renderedLine.lineGeometry.geometryId.specId).toEqual(SPEC_ID);
+      expect(renderedLine.lineGeometry.transform).toEqual({ x: 0, y: 0 });
     });
-    test('Can render points points', () => {
+    test('Can render points', () => {
       const {
-        areaGeometry: { points },
+        lineGeometry: { points },
         indexedGeometries,
-      } = renderedArea;
+      } = renderedLine;
       // all the points minus the undefined ones on a log scale
       expect(points.length).toBe(7);
       // all the points expect null geometries
@@ -921,7 +908,7 @@ describe('Rendering points - areas', () => {
       expect(zeroValueIndexdGeometry).toBeDefined();
       expect(zeroValueIndexdGeometry.length).toBe(1);
       // moved to the bottom of the chart
-      expect(zeroValueIndexdGeometry[0].y).toBe(100);
+      expect((zeroValueIndexdGeometry[0] as PointGeometry).y).toBe(100);
       // 0 radius point
       expect((zeroValueIndexdGeometry[0] as PointGeometry).radius).toBe(0);
     });
