+++ src/types/xy_chart/rendering/rendering.test.ts
@@ -1,5 +1,5 @@
-import { DEFAULT_GEOMETRY_STYLES } from '../../utils/themes/theme_commons';
-import { getSpecId } from '../../utils/ids';
+import { DEFAULT_GEOMETRY_STYLES } from '../../../utils/themes/theme_commons';
+import { getSpecId } from '../../../utils/ids';
 import { BarGeometry, getGeometryStyle, isPointOnGeometry, PointGeometry } from './rendering';
 
 describe('Rendering utils', () => {
